+++ src/Framework/Bootstrap/GacelaConfig.php
@@ -6,8 +6,8 @@ namespace Gacela\Framework\Bootstrap;
 
 use Closure;
 use Gacela\Framework\Config\ConfigReaderInterface;
+use Gacela\Framework\Config\GacelaConfigBuilder\BindingsBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\ConfigBuilder;
-use Gacela\Framework\Config\GacelaConfigBuilder\MappingInterfacesBuilder;
 use Gacela\Framework\Config\GacelaConfigBuilder\SuffixTypesBuilder;
 use Gacela\Framework\Event\GacelaEventInterface;
 use Gacela\Framework\Plugin\PluginInterface;
@@ -18,7 +18,7 @@ final class GacelaConfig
 
     private SuffixTypesBuilder $suffixTypesBuilder;
 
-    private MappingInterfacesBuilder $mappingInterfacesBuilder;
+    private BindingsBuilder $bindingBuilder;
 
     /** @var array<string, class-string|object|callable> */
     private array $externalServices;
@@ -57,7 +57,7 @@ final class GacelaConfig
         $this->externalServices = $externalServices;
         $this->configBuilder = new ConfigBuilder();
         $this->suffixTypesBuilder = new SuffixTypesBuilder();
-        $this->mappingInterfacesBuilder = new MappingInterfacesBuilder();
+        $this->bindingBuilder = new BindingsBuilder();
     }
 
     /**
@@ -137,14 +137,26 @@ final class GacelaConfig
     }
 
     /**
+     * @deprecated in favor of `$this->addBinding()`
+     * It will be removed in the next release
+     *
+     * @param class-string $key
+     * @param class-string|object|callable $value
+     */
+    public function addMappingInterface(string $key, string|object|callable $value): self
+    {
+        return $this->addBinding($key, $value);
+    }
+
+    /**
      * Define the mapping between interfaces and concretions, so Gacela services will auto-resolve them automatically.
      *
      * @param class-string $key
      * @param class-string|object|callable $value
      */
-    public function addMappingInterface(string $key, $value): self
+    public function addBinding(string $key, string|object|callable $value): self
     {
-        $this->mappingInterfacesBuilder->bind($key, $value);
+        $this->bindingBuilder->bind($key, $value);
 
         return $this;
     }
@@ -330,7 +342,7 @@ final class GacelaConfig
      *     external-services: array<string,class-string|object|callable>,
      *     config-builder: ConfigBuilder,
      *     suffix-types-builder: SuffixTypesBuilder,
-     *     mapping-interfaces-builder: MappingInterfacesBuilder,
+     *     mapping-interfaces-builder: BindingsBuilder,
      *     should-reset-in-memory-cache: ?bool,
      *     file-cache-enabled: ?bool,
      *     file-cache-directory: ?string,
@@ -351,7 +363,7 @@ final class GacelaConfig
             'external-services' => $this->externalServices,
             'config-builder' => $this->configBuilder,
             'suffix-types-builder' => $this->suffixTypesBuilder,
-            'mapping-interfaces-builder' => $this->mappingInterfacesBuilder,
+            'mapping-interfaces-builder' => $this->bindingBuilder,
             'should-reset-in-memory-cache' => $this->shouldResetInMemoryCache,
             'file-cache-enabled' => $this->fileCacheEnabled,
             'file-cache-directory' => $this->fileCacheDirectory,